carvi_image_prediction/requirements.txt
Django==4.1.7
pymongo==3.12 ## Super importante este para evitar un erro que genera los modelos de Djongo al llamar get, create, etc
djangorestframework==3.14.0